fix für den fix
Former-commit-id: 704e4d989bf9954b6b90cfcc5be5be1bb8b6d001

+++ src/fitdata.m
@@ -30,8 +30,8 @@ function [params, p_err, chisq] = fitdata(model, x, y, err, start_in, fix)
     fo = fitoptions('Method', 'NonlinearLeastSquares', 'lower', lbound,...
                     'upper', ubound, 'weights', 1./err,... % <- 1./err is important!
                     'StartPoint', start);
-                            
-    [fitobject, gof] = fit(x, y, ft, fo);
+
+    [fitobject, gof] = fit(double(x), y, ft, fo);
     
     params = zeros(length(start_in),1);
     params(fix_ind) = start_in(fix_ind);
